requirements.txt
# Core
openai>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.66.0
google>=3.0.0

